# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Administrative REST API for managing the MCP Data Platform.\nEndpoints cover system info, configuration, personas, auth keys, audit logs, and knowledge management.",
        "title": "MCP Data Platform Admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/admin",
    "paths": {
...
        "/personas": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all configured personas with tool counts.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Personas"
                ],
                "summary": "List personas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.personaListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new persona. Only available in database config mode.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Personas"
                ],
                "summary": "Create persona",
                "parameters": [
                    {
                        "description": "Persona definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.personaCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/admin.personaDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            }
        },
        "/personas/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single persona with resolved tool list.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Personas"
                ],
                "summary": "Get persona",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Persona name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.personaDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing persona. Only available in database config mode.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Personas"
                ],
                "summary": "Update persona",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Persona name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Persona definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.personaCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.personaDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a persona. Only available in database config mode. Cannot delete the admin persona.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Personas"
                ],
                "summary": "Delete persona",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Persona name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.statusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.personaContextDetail": {
            "type": "object",
            "properties": {
                "agent_instructions_override": {
                    "type": "string"
                },
                "agent_instructions_suffix": {
                    "type": "string",
                    "example": "Focus on schema details, data lineage, and query optimization."
                },
                "description_override": {
                    "type": "string"
                },
                "description_prefix": {
                    "type": "string",
                    "example": "This user is a data engineer responsible for ETL pipelines."
                }
            }
        },
        "admin.personaCreateRequest": {
            "type": "object",
            "properties": {
                "agent_instructions_override": {
                    "type": "string"
                },
                "agent_instructions_suffix": {
                    "type": "string"
                },
                "allow_connections": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allow_tools": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "datahub_*"
                    ]
                },
                "deny_connections": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "deny_tools": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string",
                    "example": "Read-only access to DataHub"
                },
                "description_override": {
                    "type": "string"
                },
                "description_prefix": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string",
                    "example": "Data Viewer"
                },
                "name": {
                    "type": "string",
                    "example": "viewer"
                },
                "priority": {
                    "type": "integer",
                    "example": 0
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "viewer"
                    ]
                }
            }
        },
        "admin.personaDetail": {
            "type": "object",
            "properties": {
                "allow_connections": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allow_tools": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "trino_*",
                        "datahub_*",
                        "s3_*"
                    ]
                },
                "context": {
                    "$ref": "#/definitions/admin.personaContextDetail"
                },
                "deny_connections": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "deny_tools": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "capture_insight"
                    ]
                },
                "description": {
                    "type": "string",
                    "example": "ETL pipeline development and schema management"
                },
                "display_name": {
                    "type": "string",
                    "example": "Data Engineer"
                },
                "name": {
                    "type": "string",
                    "example": "data-engineer"
                },
                "priority": {
                    "type": "integer",
                    "example": 10
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "data_engineer"
                    ]
                },
                "source": {
                    "description": "\"file\", \"database\", or \"both\"",
                    "type": "string",
                    "example": "file"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "trino_query",
                        "trino_describe_table",
                        "datahub_search"
                    ]
                }
            }
        },
        "admin.personaListResponse": {
            "type": "object",
            "properties": {
                "personas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin.personaSummary"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 6
                }
            }
        },
        "admin.personaSummary": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "ETL pipeline development and schema management"
                },
                "display_name": {
                    "type": "string",
                    "example": "Data Engineer"
                },
                "name": {
                    "type": "string",
                    "example": "data-engineer"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "data_engineer"
                    ]
                },
                "source": {
                    "description": "\"file\", \"database\", or \"both\"",
                    "type": "string",
                    "example": "file"
                },
                "tool_count": {
                    "type": "integer",
                    "example": 19
                }
            }
        },
        "admin.problemDetail": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "resource not found"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                },
                "title": {
                    "type": "string",
                    "example": "Not Found"
                },
                "type": {
                    "type": "string",
                    "example": "about:blank"
                }
            }
        },
...
        "admin.statusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}